requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
clickhouse-connect==0.14.1
exceptiongroup==1.3.1
fastapi==0.135.3
h11==0.16.0
idna==3.11
joblib==1.5.3
lz4==4.4.5
numpy==2.2.6
nvidia-nccl-cu12==2.29.7
pandas==2.3.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.44.0
xgboost==3.2.0
zstandard==0.25.0
